refactor: simplify ScreenshotBranding component by removing iOS-specific logic
- Removed the implementation details for rendering the ScreenshotBranding component on iOS Notch devices.
- The component now returns null, effectively disabling its functionality while retaining the structure for future updates.
